Flatten redundant modules in parser and morphology
- Move `check_recursion_depth` from `src/parser/recursion.rs` to `src/parser/common.rs`.
- Move `match_suffix` from `src/morphology/matcher.rs` to `src/morphology/mod.rs`.
- Delete `src/parser/recursion.rs` and `src/morphology/matcher.rs`.
- Update imports and tests.
- Document changes in `.jules/razor.md`.

This reduces file count and module depth, adhering to the Razor philosophy of simplicity.
